refactor: domain 이름 변경
