+++ src/main/java/org/example/autoreview/domain/notification/domain/NotificationRepository.java
@@ -1,6 +1,6 @@
-package org.example.autoreview.domain.scheduler.domain;
+package org.example.autoreview.domain.notification.domain;
 
-import org.example.autoreview.domain.scheduler.enums.NotificationStatus;
+import org.example.autoreview.domain.notification.enums.NotificationStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
